% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/status.R
\name{autostatus}
\alias{autostatus}
\title{Auto-find logfiles and show progress}
\usage{
autostatus(path = NULL, here = TRUE, s = 0.1)
}
\arguments{
\item{path}{path to project root}

\item{here}{should base path of the logfile be located using `here::here()`}

\item{s}{update interval in seconds}
}
\value{
nothing, invoked for side effects
}
\description{
This is a wrapper around `logvogel::status()`
}
\examples{
\dontrun{
autostatus()
}
}
