% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logfile.R
\name{logfile}
\alias{logfile}
\title{Logfile for external progress tracking}
\usage{
logfile(path, n, here = TRUE)
}
\arguments{
\item{path}{path to the logfile}

\item{n}{total number of iterations}

\item{here}{should base path of the logfile be located using `here::here()`}
}
\value{
A ref class with methods `update()`, `print()`, and `remove()`
}
\description{
This reference class keeps track of a logfile, which can be updated within a
loop etc. and tracked from _outside_ (in an external process), using
`status()`.
}
\examples{
\dontrun{
log <- logfile("logfile.log", 100)
}
}
\keyword{internal}
