% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/status.R
\name{status}
\alias{status}
\title{Show progress using a logfile}
\usage{
status(path, here = TRUE, s = 0.1)
}
\arguments{
\item{path}{path to the logfile}

\item{here}{should base path of the logfile be located using `here::here()`}

\item{s}{update interval in seconds}
}
\value{
nothing, invoked for side effects
}
\description{
Use this function in an external R process to track the status of a running
loop from logfile.
}
\examples{
\dontrun{
status("logfile.log")
}
}
